# Minor: include the number of files run in sqllogictest display