## Which issue does this PR close?
- Follow on to https://github.com/apache/datafusion/pull/14355
- Follow on to https://github.com/apache/datafusion/pull/13936

## Rationale for this change

While testing https://github.com/apache/datafusion/pull/14355 from @findepi  I found it hard to see how many files were run

## What changes are included in this PR?

Before this PR:
```shell
andrewlamb@Andrews-MacBook-Pro-2:~/Software/datafusion$ cargo test --test sqllogictests  -- union
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.16s
     Running bin/sqllogictests.rs (target/debug/deps/sqllogictests-3d953efd822e1e41)
Completed in 0 seconds
```

After this PR, it reports `Completed 2 tests`:

```shell
andrewlamb@Andrews-MacBook-Pro-2:~/Software/datafusion$ cargo test --test sqllogictests  -- union
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.16s
     Running bin/sqllogictests.rs (target/debug/deps/sqllogictests-3d953efd822e1e41)
Completed 2 tests in 0 seconds
```

## Are these changes tested?

I tested them manually
## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->